requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-axes==6.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
drf-yasg==1.21.7
requests==2.31.0
argon2-cffi==23.1.0
phonenumbers==8.13.27
django-phonenumber-field==7.2.0
django-filter==23.5
setuptools  # Added for pkg_resources
Pillow  # Added for ImageField